requirements.txt
crewai>=0.18.7
langchain-community>=0.0.8
# Replace OpenAI with Ollama integration
langchain-ollama>=0.0.1
litellm>=0.9.4
matplotlib>=3.8.0
numpy>=1.24.3
pandas>=2.0.3
python-dotenv>=1.0.0
requests>=2.31.0

# LangChain packages for model integration
langchain-core>=0.1.10

# Data visualization and reporting
openpyxl>=3.1.0
nbformat>=5.7.0
tabulate>=0.9.0

# For Ollama model support (Llama 3.3 70B)
ollama>=0.1.5

# Optional: For alternative local models
# gpt4all>=2.0.2 